requirements.txt
django >= 3.0.6
Pillow >= 7.1.2
httpagentparser >= 1.9.0